Look back 1 year by default for log display with only manual logs
Also localizes the PeriodFilterControl and adds 1m and 1y options
